api2/requirements.txt
fastapi
Pillow
torch
transformers